fix marker styling on journey diagram
