+++ src/diagrams/user-journey/journeyRenderer.js
@@ -3,6 +3,7 @@ import svgDraw from './svgDraw';
 import { getConfig } from '../../config';
 import { configureSvgSize } from '../../setupGraphViewbox';
 import addSVGAccessibilityFields from '../../accessibility';
+import { markerUrl } from '../../utils';
 
 export const setConf = function (cnf) {
   const keys = Object.keys(cnf);
@@ -112,7 +113,7 @@ export const draw = function (text, id, version, diagObj) {
     .attr('y2', conf.height * 4)
     .attr('stroke-width', 4)
     .attr('stroke', 'black')
-    .attr('marker-end', 'url(#arrowhead)');
+    .attr('marker-end', markerUrl(diagram, 'arrowhead'));
 
   const extraVertForTitle = title ? 70 : 0;
   diagram.attr('viewBox', `${box.startx} -25 ${width} ${height + extraVertForTitle}`);
